pyproject.toml
[project]
name = "opendic-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "akongstad", email = "kongstad@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "snowflake-connector-python>=3.14.1",
]

[build-system]
requires = ["uv_build>=0.6.12,<0.7"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
